Assignment3/sentiment-analysis-with-cloud-services-master/requirements.txt
watson_developer_cloud
google-cloud-language
boto3
pandas